move storage and limitedplacing mechani into gameplay
