+++ src/main/java/io/th0rgal/oraxen/mechanics/provided/gameplay/storage/StorageMechanic.java
@@ -1,4 +1,4 @@
-package io.th0rgal.oraxen.utils.storage;
+package io.th0rgal.oraxen.mechanics.provided.gameplay.storage;
 
 import com.jeff_media.morepersistentdatatypes.DataType;
 import com.ticxo.modelengine.api.ModelEngineAPI;
